# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Past Questions Practice Website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      tags: [Subjects]
      operationId: listSubjects
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"

    post:
      tags: [Subjects]
      operationId: createSubject
      summary: Create a subject (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    patch:
      tags: [Subjects]
      operationId: updateSubject
      summary: Update a subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"

    delete:
      tags: [Subjects]
      operationId: deleteSubject
      summary: Delete a subject (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Exam Types ──────────────────────────────────────────────────────────────
...
  /questions:
    get:
      tags: [Questions]
      operationId: listQuestions
      summary: List questions with optional filters
      parameters:
        - name: examTypeId
          in: query
          required: false
          schema: { type: integer }
        - name: subjectId
          in: query
          required: false
          schema: { type: integer }
        - name: year
          in: query
          required: false
          schema: { type: integer }
        - name: topic
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50 }
        - name: published
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionListResponse"

    post:
      tags: [Questions]
      operationId: createQuestion
      summary: Create a question (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{questionId}:
    get:
      tags: [Questions]
      operationId: getQuestion
      summary: Get a question by ID
      parameters:
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found

    patch:
      tags: [Questions]
      operationId: updateQuestion
      summary: Update a question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

    delete:
      tags: [Questions]
      operationId: deleteQuestion
      summary: Delete a question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /questions/{questionId}/publish:
    patch:
      tags: [Questions]
      operationId: publishQuestion
      summary: Publish or unpublish a question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/years:
    get:
      tags: [Questions]
      operationId: listAvailableYears
      summary: List available years for a given exam type and subject
      parameters:
        - name: examTypeId
          in: query
          required: false
          schema: { type: integer }
        - name: subjectId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of years
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer

  # ─── Exam Sessions (Practice Sets) ──────────────────────────────────────────
...
components:
...
  schemas:
...
    Subject:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    SubjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }

    # ── Exam Types ────────────────────────────────────────────────────────────
...
    AnswerOption:
      type: object
      required: [id, letter, text]
      properties:
        id: { type: integer }
        letter: { type: string }
        text: { type: string }
    Question:
      type: object
      required: [id, text, examTypeId, subjectId, year, options, correctLetter, published, createdAt]
      properties:
        id: { type: integer }
        text: { type: string }
        examTypeId: { type: integer }
        subjectId: { type: integer }
        examTypeName: { type: ["string", "null"] }
        subjectName: { type: ["string", "null"] }
        year: { type: integer }
        topic: { type: ["string", "null"] }
        explanation: { type: ["string", "null"] }
        options: { type: array, items: { $ref: "#/components/schemas/AnswerOption" } }
        correctLetter: { type: string }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
    QuestionInput:
      type: object
      required: [text, examTypeId, subjectId, year, options, correctLetter]
      properties:
        text: { type: string, minLength: 1 }
        examTypeId: { type: integer }
        subjectId: { type: integer }
        year: { type: integer }
        topic: { type: string }
        explanation: { type: string }
        correctLetter: { type: string, minLength: 1 }
        published: { type: boolean, default: false }
        options:
          type: array
          minItems: 2
          items:
            type: object
            required: [letter, text]
            properties:
              letter: { type: string }
              text: { type: string }
    QuestionUpdate:
      type: object
      properties:
        text: { type: string, minLength: 1 }
        examTypeId: { type: integer }
        subjectId: { type: integer }
        year: { type: integer }
        topic: { type: string }
        explanation: { type: string }
        correctLetter: { type: string }
        published: { type: boolean }
        options:
          type: array
          minItems: 2
          items:
            type: object
            required: [letter, text]
            properties:
              letter: { type: string }
              text: { type: string }
    PublishToggle:
      type: object
      required: [published]
      properties:
        published: { type: boolean }
    QuestionListResponse:
      type: object
      required: [questions, total, page, limit]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Exam Sessions ──────────────────────────────────────────────────────────
...