# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Past Questions Practice Website API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionUpdateOptionsItem } from './questionUpdateOptionsItem';

export interface QuestionUpdate {
  /** @minLength 1 */
  text?: string;
  examTypeId?: number;
  subjectId?: number;
  year?: number;
  topic?: string;
  explanation?: string;
  correctLetter?: string;
  published?: boolean;
  /** @minItems 2 */
  options?: QuestionUpdateOptionsItem[];
}